+++ tests/testAlfil.py
@@ -1,4 +1,9 @@
-""""import unittest
+import sys
+import os
+
+sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
+
+import unittest
 from chess.bishop import Bishop
 from chess.board import Board
 
@@ -80,4 +85,3 @@ class TestBishop(unittest.TestCase):
 
 if __name__ == '__main__':
     unittest.main()
-"""