+++ tests/testCaballo.py
@@ -1,4 +1,9 @@
-"""
+import sys
+import os
+
+sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
+
+
 import unittest
 from chess.knight import Knight
 from chess.board import Board
@@ -52,4 +57,3 @@ class TestKnight(unittest.TestCase):
 
 if __name__ == '__main__':
     unittest.main()
-"""