+++ tests/testExcepciones.py
@@ -1,3 +1,9 @@
+import sys
+import os
+
+sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
+
+
 import unittest
 from chess.exceptions import InvalidMove
 from chess.exceptions import InvalidTurn