+++ tests/testPeon.py
@@ -1,4 +1,9 @@
-"""
+"""import sys
+import os
+
+sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
+
+
 import unittest
 from chess.pawn import Pawn
 from chess.board import Board
@@ -65,5 +70,4 @@ class TestPawn(unittest.TestCase):
 
 
 if __name__ == '__main__':
-    unittest.main()
-"""
+    unittest.main()"""