+++ tests/testPiezas.py
@@ -1,4 +1,9 @@
-"""
+import sys
+import os
+
+sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
+
+
 import unittest
 from chess.pieces import Piece
 from chess.board import Board
@@ -71,4 +76,3 @@ class TestPieces(unittest.TestCase):
 
 if __name__ == '__main__':
     unittest.main()
-"""