+++ tests/testReina.py
@@ -1,4 +1,9 @@
-"""
+import sys
+import os
+
+sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
+
+
 import unittest
 from chess.queen import Queen
 from chess.board import Board
@@ -46,4 +51,3 @@ class TestQueenValidPositions(unittest.TestCase):
 
 if __name__ == "__main__":
     unittest.main()
-"""