+++ tests/testRey.py
@@ -1,4 +1,9 @@
-"""
+import sys
+import os
+
+sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
+
+
 import unittest
 from chess.king import King
 from chess.board import Board
@@ -84,4 +89,3 @@ class TestKing(unittest.TestCase):
 
 if __name__ == '__main__':
     unittest.main()
-"""