+++ tests/testTorres.py
@@ -1,4 +1,9 @@
-"""
+import sys
+import os
+
+sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
+
+
 import unittest
 from chess.rook import Rook
 from chess.board import Board
@@ -59,4 +64,3 @@ class TestRook(unittest.TestCase):
 
 if __name__ == '__main__':
     unittest.main()
-"""